refactor(multiple): ♻️ move dislocated libs to `libs` folder from root folder
