+++ libs/matches-overview/project.json
@@ -1,7 +1,7 @@
 {
   "name": "matches-overview",
-  "$schema": "../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "matches-overview/src",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/matches-overview/src",
   "prefix": "lib",
   "projectType": "library",
   "tags": [],
@@ -10,7 +10,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "matches-overview/jest.config.ts"
+        "jestConfig": "libs/matches-overview/jest.config.ts"
       }
     },
     "lint": {
