+++ libs/tag-input/jest.config.ts
@@ -1,8 +1,8 @@
 export default {
-  displayName: 'matches-overview',
-  preset: '../jest.preset.js',
+  displayName: 'tag-input',
+  preset: '../../jest.preset.js',
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-  coverageDirectory: '../coverage/matches-overview',
+  coverageDirectory: '../../coverage/libs/tag-input',
   transform: {
     '^.+\\.(ts|mjs|js|html)$': [
       'jest-preset-angular',
