+++ libs/tag-input/tsconfig.json
@@ -18,7 +18,7 @@
       "path": "./tsconfig.spec.json"
     }
   ],
-  "extends": "../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
     "strictInjectionParameters": true,
