+++ tsconfig.base.json
@@ -17,10 +17,10 @@
     "paths": {
       "@portfolio/home": ["libs/home/src/index.ts"],
       "@portfolio/legal": ["libs/legal/src/index.ts"],
-      "@portfolio/matches-overview": ["matches-overview/src/index.ts"],
+      "@portfolio/matches-overview": ["libs/matches-overview/src/index.ts"],
       "@portfolio/projects": ["libs/projects/src/index.ts"],
       "@portfolio/shared-ui": ["libs/shared-ui/src/index.ts"],
-      "@portfolio/tag-input": ["tag-input/src/index.ts"]
+      "@portfolio/tag-input": ["libs/tag-input/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
